Rewrite event handling and add support for mouse events
Currently support for keydown, keyup, mousedown, mouseup and mousemotion
events.

The new implementation is not pretty, but trust me, it works.
SDL_Event is a union of C-structs. dotnet is not keen on letting
us write a "union struct" that does not necessarily have the same alignment of
fields in each possible member-struct.

By "chunking through it" and manually converting to records all of this
is avoided.